Reorganize passive login into its own module
