+++ app/controllers/application_controller.rb
@@ -5,9 +5,7 @@ class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
 
   layout Proc.new { |controller| (controller.request.xhr?) ? false : "login" }
-
-  PASSIVE_SHIBBOLETH_URL_STRING = "/Shibboleth.sso/Login?isPassive=true&target="
-  SHIBBOLETH_COOKIE_PATTERN = "_shibsession_"
+  
   LOGGED_IN_COOKIE_NAME = '_nyulibraries_logged_in'
   ESHELF_COOKIE_NAME = '_nyulibraries_eshelf_passthru'
